requirements.txt
langchain
openai
faiss-cpu
tqdm
streamlit
python-dotenv
PyPDF2
